idea_cms_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
coverage==7.3.2
factory-boy==3.3.0
pytest-django==4.7.0
black==23.10.1
flake8==6.1.0
isort==5.12.0


# إضافات جديدة للمصادقة وإدارة المهام
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0

